# io: optimizing the chances of large write in copy_bidirectional and copy