Although this PR was originally related to https://github.com/tokio-rs/tokio/issues/6519#issue-2267067024, it aims now to improve a bit the chances of large write in `io::copy` and `io::copy_bidirectional`.

## Motivation

`tokio::io::copy_bidirectional` and `tokio::io::copy` can, in some circumstances, lead to some non-optimal chances of large write when both writing and reading are pending ([see here](https://github.com/tokio-rs/tokio/blob/master/tokio/src/io/util/copy.rs#L64-L68)). For more information, refer to the issue linked above.

This PR aims to improve this behavior by reading more often.

## Solution

### Changelog

- Try to read when the buffer is not full
- Reset partially the buffer state as soon as data is fully written

### Benchmarks

```shell
copy_mem_to_mem         time:   [131.88 µs 133.09 µs 134.48 µs]
                        change: [-4.3908% -2.6619% -1.0456%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

copy_mem_to_slow_hdd    time:   [181.26 ms 181.30 ms 181.35 ms]
                        change: [-0.0183% +0.0132% +0.0477%] (p = 0.44 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

copy_chunk_to_mem       time:   [128.11 ms 128.13 ms 128.16 ms]
                        change: [-0.0458% -0.0191% +0.0075%] (p = 0.16 > 0.05)
                        No change in performance detected.

copy_chunk_to_slow_hdd  time:   [141.29 ms 141.43 ms 141.59 ms]
                        change: [-5.1731% -4.7362% -4.2937%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) high mild
  9 (9.00%) high severe
```