@mox692 2024-05-04
I think this change could make the benchmark worse in cases where the `poll_flush` takes a long time, because now the `poll_write` is always executed after the flush finishes in each loop.

I tested using `copy` benchmark:


```bash
$ cd benches && cargo bench --bench copy


copy_mem_to_slow_hdd    time:   [767.41 ms 770.31 ms 773.42 ms]
                        change: [+294.24% +298.58% +302.79%] (p = 0.00 < 0.05)
                        Performance has regressed.

copy_chunk_to_slow_hdd  time:   [789.88 ms 792.86 ms 795.96 ms]
                        change: [+425.55% +433.04% +440.10%] (p = 0.00 < 0.05)
                        Performance has regressed.
```