@mox692 2024-05-04
My question here is, in `AsyncWrite`, should `poll_write` not be executed while `poll_flush` is pending?

If so, I think this change is acceptable even if it will worsen performance, because this PR is simply trying to do right thing in that assumption.