@Armillus 2024-05-04
> > Overall performance seemed to improve a bit with the new read condition
> 
> I think this change could make the benchmark worse in cases where the `poll_flush` takes a long time, because now the `poll_write` is always executed after the flush finishes in each loop.
> 
> I tested using `copy` benchmark:
> 
> ```shell
> $ cd benches && cargo bench --bench copy
> 
> 
> copy_mem_to_slow_hdd    time:   [767.41 ms 770.31 ms 773.42 ms]
>                         change: [+294.24% +298.58% +302.79%] (p = 0.00 < 0.05)
>                         Performance has regressed.
> 
> copy_chunk_to_slow_hdd  time:   [789.88 ms 792.86 ms 795.96 ms]
>                         change: [+425.55% +433.04% +440.10%] (p = 0.00 < 0.05)
>                         Performance has regressed.
> ```

Thank you for pointing this, I've missed this benchmark when looking for it. This huge regression was due to the fact that we tried to flush even when no flush was issued during a pending read. Hence, I've implemented a [few changes](5386820ef594e2678aa9778b6d58a6f435e207a0) that change this behavior and thus reduce this huge performance regression to something more acceptable at the cost of an additional boolean in `CopyBuffer`:

```shell
copy_mem_to_mem         time:   [159.04 µs 161.96 µs 165.48 µs]
                        change: [-2.3046% +1.7540% +5.9405%] (p = 0.40 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe

copy_mem_to_slow_hdd    time:   [181.27 ms 181.42 ms 181.60 ms]
                        change: [+0.0776% +0.1623% +0.2530%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 14 outliers among 100 measurements (14.00%)
  14 (14.00%) high severe

copy_chunk_to_mem       time:   [128.06 ms 128.10 ms 128.15 ms]
                        change: [-0.1713% -0.0813% -0.0033%] (p = 0.05 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

copy_chunk_to_slow_hdd  time:   [176.05 ms 176.75 ms 177.42 ms]
                        change: [+16.939% +17.675% +18.433%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) low mild
```

As a nice side effect, it makes the flushing condition more readable in my opinion.