@Armillus 2024-05-04
> My question here is, in `AsyncWrite`, should `poll_write` not be executed while `poll_flush` is pending?

That's my understanding, or it is implementation-defined. Either way, I think it should probably be documented if that's possible.

What is sure is that the flushing future needs to be properly polled until completion, which was not the case before. With the new changes, I think that the small performance regression is not a big deal anymore (I may be wrong though).

However, note that if we consider that `poll_write()` can be executed while `poll_flush` is pending, we can likely optimize the code further with the new boolean member `flush_in_progress`.