@Darksonn 2024-05-04
I can't completely tell what's going on here, it but it sounds like you `AsyncWrite` is incorrectly implemented. Starting a flush with `poll_flush` and then changing your mind to write more data with `poll_write`, even though your last `poll_flush` returned pending, is allowed.

If your `AsyncWrite` doesn't support that, then you may need to have your `poll_write` method wait for the current flush to finish before it accepts any writes.